fix recursive dependency issue and some of the unit tests
