+++ src/vs/editor/common/codecs/frontMatterCodec/tokens/frontMatterRecord.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { assert } from '../../../../../base/common/assert.js';
 import { Colon, Word, Dash, Space, Tab } from '../../simpleCodec/tokens/index.js';
 import { FrontMatterToken, FrontMatterValueToken, TValueTypeName } from '../tokens/frontMatterToken.js';
 
@@ -92,46 +91,6 @@ export class FrontMatterRecord extends FrontMatterToken<readonly [FrontMatterRec
 	 */
 	public get valueToken(): FrontMatterValueToken<TValueTypeName> {
 		return this.tokens[2];
-	}
-
-	/**
-	 * Create new instance from a list of tokens.
-	 *
-	 * @throws if:
-	 *  - the list of tokens is not exactly 3 tokens long
-	 * 	- the first token in the list is not a `FrontMatterRecordName`
-	 * 	- the second token in the list is not a `FrontMatterRecordDelimiter`
-	 * 	- the third token in the list is not a `FrontMatterValueToken`
-	 *
-	 */
-	public static fromTokens(
-		tokens: readonly FrontMatterToken[],
-	): FrontMatterRecord {
-		assert(
-			tokens.length === 3,
-			`A front matter record must consist of exactly 3 tokens, got '${tokens.length}'.`,
-		);
-
-		const token1 = tokens[0];
-		const token2 = tokens[1];
-		const token3 = tokens[2];
-
-		assert(
-			token1 instanceof FrontMatterRecordName,
-			`Token #1 must be a front matter record name, got '${token1}'.`,
-		);
-		assert(
-			token2 instanceof FrontMatterRecordDelimiter,
-			`Token #2 must be a front matter record delimiter, got '${token2}'.`,
-		);
-		assert(
-			token3 instanceof FrontMatterValueToken,
-			`Token #3 must be a front matter value, got '${token3}'.`,
-		);
-
-		return new FrontMatterRecord([
-			token1, token2, token3,
-		]);
 	}
 
 	public override toString(): string {
